Move MongoControllerHelpers to a separate file
